requirements/base.txt
base58==0.2.4
boto==2.43.0
boto3==1.4.1
botocore==1.4.67
click==6.6
Django==1.10.2
django-appconf==1.0.2
django-bootstrap3==7.1.0
django-compressor==2.1
django-storages==1.5.1
djangorestframework==3.5.1
docutils==0.12
futures==3.0.5
hjson==2.0.2
jmespath==0.9.0
kappa==0.6.0
lambda-packages==0.10.0
placebo==0.8.1
python-dateutil==2.5.3
python-slugify==1.2.1
PyYAML==3.12
rcssmin==1.0.6
requests==2.11.1
rjsmin==1.0.12
s3transfer==0.1.9
six==1.10.0
tqdm==4.9.0
troposphere==1.8.2
Unidecode==0.4.19
Werkzeug==0.11.11
wsgi-request-logger==0.4.6
zappa==0.28.1
